requirements.txt
streamlit>=1.16
pyyaml
numpy<1.24
pandas
matplotlib
plotly
bokeh
hvplot
geopandas
networkx
holoviews
panel
geoviews
cartopy
pyproj
scipy
xarray
netcdf4
pypsa